[server/mobileRefectoring] "../vending/delete" 리펙토링
1. 설명 주석 추가
2. 클라이언트 요청 데이터 검사 구문 추가
3. 클라이언트 요청 데이터가 없거나 데이터 삭제시 에러날 경우 "false"와 에러 메세지(msg) 응답

+++ server/routes/mobile.js
@@ -118,18 +118,23 @@ router.post('/vending/update', (req, res) => {
     const vending = req.body.vending;
 
     // 클라이언트가 요청한 데이터가 있는지 검사
-    if (!String.isEmpty(userId)) {
+    if (!String.isEmpty(vending)) {
         // 클라이언트가 전송한 "vending" 이 있다면, 자판기 정보 수정
         db.query(`UPDATE vendings SET vending_name=?, vending_description=?, vending_full_size=? WHERE serial_number=?;`, 
             [vending.name, vending.description, vending.fullSize, vending.serialNumber], (err, result) => {
                 // 실패시 false 응답
                 if (err) {
                     console.log(err);
-                    res.json({ success : false });
+                    res.json({ 
+                        success : false,
+                        msg : err
+                    });
                 }
 
                 // 성공시 true 응답
-                res.json({ success : true });
+                res.json({ 
+                    success : true 
+                });
         });
     } else {
         // 클라이언트가 전송한 데이터가 없다면 false 반환
@@ -142,18 +147,34 @@ router.post('/vending/update', (req, res) => {
 
 // 자판기 삭제 요청
 router.post('/vending/delete', (req, res) => {
+    // 클라이언트가 요청한 데이터 저장
     const serialNumber = req.body.serialNumber;
 
-    db.query(`DELETE FROM vendings WHERE serial_number=?;`, [serialNumber], (err, result) => {
-        // 실패시 false 응답
-        if (err) {
-            console.log(err);
-            res.json({ success : false });
-        }
+    // 클라이언트가 요청한 데이터가 있는지 검사
+    if (!String.isEmpty(serialNumber)) {
+        // 클라이언트가 전송한 "serialNumber" 가 있다면, 자판기 정보 수정
+        db.query(`DELETE FROM vendings WHERE serial_number=?;`, [serialNumber], (err, result) => {
+            // 실패시 false 응답
+            if (err) {
+                console.log(err);
+                res.json({
+                    success : false,
+                    msg : err
+                });
+            }
 
-        // 성공시 true 응답
-        res.json({ success : true });
-    })
+            // 성공시 true 응답
+            res.json({ 
+                success : true 
+            });
+        });
+    } else {
+        // 클라이언트가 전송한 데이터가 없다면 false 반환
+        res.json({
+            success : false,
+            msg : "The serial number of the server is empty."
+        });
+    }
 });
 
 // 모듈 내보내기
